+++ src/internal.h
@@ -1,0 +1,32 @@
+#ifndef __udx__internal_h__
+#define __udx__internal_h__
+
+#include "../include/udx.h"
+
+#define UDX_PACKET_CALLBACK     (UDX_PACKET_STREAM_SEND | UDX_PACKET_STREAM_DESTROY | UDX_PACKET_SEND)
+#define UDX_PACKET_FREE_ON_SEND (UDX_PACKET_STREAM_STATE | UDX_PACKET_STREAM_DESTROY)
+
+inline void
+addr_to_v6 (struct sockaddr_in *addr) {
+  struct sockaddr_in6 in;
+  memset(&in, 0, sizeof(in));
+
+  in.sin6_family = AF_INET6;
+  in.sin6_port = addr->sin_port;
+#ifdef SIN6_LEN
+  in.sin6_len = sizeof(struct sockaddr_in6);
+#endif
+
+  in.sin6_addr.s6_addr[10] = 0xff;
+  in.sin6_addr.s6_addr[11] = 0xff;
+
+  // Copy the IPv4 address to the last 4 bytes of the IPv6 address.
+  memcpy(&(in.sin6_addr.s6_addr[12]), &(addr->sin_addr), 4);
+
+  memcpy(addr, &in, sizeof(in));
+}
+
+void udx__trigger_send_callback(udx_socket_t *socket, udx_packet_t *packet);
+void udx__close_handles(udx_socket_t *socket);
+
+#endif
